Move Tabs components to atoms folder
